+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 3.1.0
+- Add `FREObject.className`
+- Add `FreObjectSwift.className`
+- Add `@dynamicMemberLookup` for Array types to `FreObjectSwift`
+- Add `[Date]` support to `FREArray`
+- Declare `FREArray` as open
+- Change: `FREArray` now uses Vector.<Type> throughout
+
 ### 3.0.0
 - Upgrade to Xcode 10.1
 - Upgrade to iOS SDK 12.1